# restler-testing/Compile/preprocessed/openapi_preprocessed.json
{"openapi":"3.0.0","info":{"title":"Cuts.ae API","version":"1.0.0","description":"API for Cuts.ae food delivery platform"},"servers":[{"url":"http://host.docker.internal:45000/api/v1"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"paths":{"/auth/login":{"post":{"summary":"Login","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}},"required":["email","password"]}}}},"responses":{"200":{"description":"Login successful"},"401":{"description":"Invalid credentials"},"400":{"description":"Bad request"}}}},"/admin/analytics":{"get":{"summary":"Get platform analytics","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Analytics data retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/restaurants":{"get":{"summary":"Get all restaurants","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Restaurants list retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/restaurants/{id}/approve":{"post":{"summary":"Approve restaurant","tags":["Admin"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Restaurant approved"},"404":{"description":"Restaurant not found"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/drivers":{"get":{"summary":"Get all drivers","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Drivers list retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/drivers/{id}/approve":{"post":{"summary":"Approve driver","tags":["Admin"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Driver approved"},"404":{"description":"Driver not found"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/invoices":{"get":{"summary":"Get all invoices","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Invoices list retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/invoices/{id}":{"get":{"summary":"Get invoice details","tags":["Admin"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Invoice details retrieved"},"404":{"description":"Invoice not found"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/invoices/generate":{"post":{"summary":"Generate invoice","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Invoice generated"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/users":{"get":{"summary":"Get all users","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Users list retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/admin/orders":{"get":{"summary":"Get all orders (admin)","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Orders list retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}},"/orders":{"get":{"summary":"Get orders for restaurant","tags":["Orders"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"restaurant_id","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Orders retrieved"},"401":{"description":"Unauthorized"},"500":{"description":"Internal server error"}}}}}}